// src/wallet/enums/wallet-events.enum.ts
export enum WalletEvents {
  CONNECTED = "wallet.connected",
  DISCONNECTED = "wallet.disconnected",
  TRANSACTION_SENT = "wallet.transaction.sent",
  TRANSACTION_REJECTED = "wallet.transaction.rejected",
  NETWORK_SWITCHED = "wallet.network.switched",
  CONNECTION_ERROR = "wallet.connection.error",
  ERROR = "wallet.error",
}